Backend/requirements.txt
langchain==0.1.6
langchain-community==0.0.19
pinecone-client==2.2.4
python-dotenv==1.0.0
streamlit==1.29.0
unstructured==0.14.10
